requirements.txt
pytest==8.0.0
playwright==1.41.0
pytest-playwright==0.4.4
allure-pytest==2.13.2
faker==22.6.0
pytest-base-url==2.1.0
python-dotenv==1.0.1
